requirements.txt
streamlit>=1.31
pydantic>=2.6
pandas>=2.0
pyyaml>=6.0
